pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dagweaver"
version = "0.1.0"
description = "Library for Apache Airflow DAG development"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "sqlglot>=24.0.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
[project.urls]
Homepage = "https://github.com/baranovss/DagWeaver"
Repository = "https://github.com/baranovss/DagWeaver"
Documentation = "https://github.com/baranovss/DagWeaver#readme"

[project.optional-dependencies]
airflow = ["apache-airflow>=2.0.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py39"
